+++ app/models/registration.ts
@@ -3,6 +3,7 @@ import DS from 'ember-data';
 import Comment from './comment';
 import Contributor from './contributor';
 import Node from './node';
+import RegistrationSchema from './registration-schema';
 import User from './user';
 
 /**
@@ -19,6 +20,8 @@ import User from './user';
 export default class Registration extends Node.extend() {
     @attr('date') dateRegistered!: Date;
     @attr('boolean') pendingRegistrationApproval!: boolean;
+    @attr('boolean') archiving!: boolean;
+    @attr('boolean') embargoed!: boolean;
     @attr('date') embargoEndDate!: Date | null;
     @attr('boolean') pendingEmbargoApproval!: boolean;
     @attr('boolean') withdrawn!: boolean;
@@ -35,8 +38,10 @@ export default class Registration extends Node.extend() {
 
     @belongsTo('node', { inverse: 'registrations' }) registeredFrom!: DS.PromiseObject<Node> & Node;
     @belongsTo('user', { inverse: null }) registeredBy!: DS.PromiseObject<User> & User;
-    @hasMany('contributor') contributors!: DS.PromiseManyArray<Contributor>;
-    @hasMany('comment') comments!: DS.PromiseManyArray<Comment>;
+    @hasMany('contributor', { inverse: 'node' }) contributors!: DS.PromiseManyArray<Contributor>;
+    @hasMany('comment', { inverse: 'node' }) comments!: DS.PromiseManyArray<Comment>;
+    @belongsTo('registration-schema', { inverse: null })
+    registrationSchema!: DS.PromiseObject<RegistrationSchema> & RegistrationSchema;
 }
 
 declare module 'ember-data' {