+++ app/resolve-guid/resolved-guid-route.ts
@@ -4,7 +4,7 @@ import { task } from 'ember-concurrency';
 
 import { ReferentModelName } from 'ember-osf-web/models/guid';
 import Ready from 'ember-osf-web/services/ready';
-import RouteContext from 'ember-osf-web/services/route-context';
+import RouteContext, { Query } from 'ember-osf-web/services/route-context';
 
 /**
  * Base class for the root-level GUID routes
@@ -14,11 +14,16 @@ export default class ResolveGuidRoute extends Route {
     @service router!: any;
     @service routeContext!: RouteContext;
 
-    resolveGuid = task(function *(this: ResolveGuidRoute, guid: string, expectedType?: ReferentModelName) {
+    resolveGuid = task(function *(
+        this: ResolveGuidRoute,
+        guid: string,
+        expectedType?: ReferentModelName,
+        query?: Query,
+    ) {
         const blocker = this.ready.getBlocker();
 
         try {
-            const model = yield this.routeContext.setGuid(guid, expectedType);
+            const model = yield this.routeContext.setGuid(guid, expectedType, query);
             blocker.done();
             return model;
         } catch (error) {