+++ app/serializers/osf-serializer.ts
@@ -163,6 +163,15 @@ export default class OsfSerializer extends JSONAPISerializer.extend({
         if (documentHash.meta) {
             documentHash.data.attributes.apiMeta = documentHash.meta;
         }
+        if (documentHash.data.relationships) {
+            documentHash.data.attributes.relatedCounts = Object.entries(documentHash.data.relationships).reduce(
+                (acc: { [k: string]: number | undefined }, [relName, rel]: [string, any]) => {
+                    acc[camelize(relName)] = rel.links ? rel.links.related.meta.count : undefined;
+                    return acc;
+                },
+                {},
+            );
+        }
         return documentHash;
     },
 
@@ -179,6 +188,9 @@ export default class OsfSerializer extends JSONAPISerializer.extend({
 }) {
     attrs = {
         links: {
+            serialize: false,
+        },
+        relatedCounts: {
             serialize: false,
         },
     };