+++ config/environment.js
@@ -208,6 +208,7 @@ module.exports = function(environment) {
                 'guid-registration.forks': 'ember_project_forks_page',
                 'guid-node.analytics.index': 'ember_project_analytics_page',
                 'guid-registration.analytics.index': 'ember_project_analytics_page',
+                'guid-node.registrations': 'ember_project_registrations_page',
             },
             navigation: {
                 institutions: 'institutions_nav_bar',