+++ lib/osf-components/addon/locales/ja/translations.ts
@@ -1,0 +1,13 @@
+/* tslint:disable:object-literal-sort-keys max-line-length */
+
+export default {
+    'osf-components': {
+        'draft-registration-card': {
+            initiated_by: 'Initiated by:',
+            started: 'Started:',
+            last_updated: 'Last updated:',
+            register: 'Register',
+            delete_draft_confirm: 'Are you sure you want to delete this draft registration?',
+        },
+    },
+};