+++ tests/integration/components/draft-registration-card/component-test.ts
@@ -1,0 +1,14 @@
+import { render } from '@ember/test-helpers';
+import { setupRenderingTest } from 'ember-qunit';
+import hbs from 'htmlbars-inline-precompile';
+import { module, test } from 'qunit';
+
+module('Integration | Component | draft-registration-card', hooks => {
+    setupRenderingTest(hooks);
+
+    test('it renders', async function(assert) {
+        await render(hbs`{{draft-registration-card}}`);
+
+        assert.dom(this.element).hasText('Edit Delete Register');
+    });
+});